requirements.txt
allure_python_commons==2.10.0
behave==1.2.6
page_objects==1.1.0
selenium==4.4.3
webdriver_manager==3.8.3
allure-behave==2.10.0